+++ nixio/pycore/__init__.py
@@ -5,4 +5,3 @@ from .dimensions import RangeDimension, SetDimension, SampledDimension
 from .feature import Feature
 from .tag import Tag
 from .multi_tag import MultiTag
-from .group import Group
